Max/rightsurvey-xp-api/requirements.txt
falcon
gunicorn==20.0.4
requests
redis
greenlet
gevent==20.6.2
falcon-swagger-ui
pymongo==3.9.0
falcon-auth==1.1.0
qrcode
jsonschema
reportlab
XlsxWriter
apscheduler
confluent-kafka==0.9.4
fluent-logger==0.9.6
falcon-multipart
simple-salesforce==1.10.1
msal==1.5.0